requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data access
wrds>=3.1.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Optimization
numba>=0.58.0

# Configuration
pyyaml>=6.0

# Broker APIs
ib-insync>=0.9.86  # Interactive Brokers
alpaca-trade-api>=3.0.0  # Alpaca

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0

# Logging and monitoring
colorlog>=6.7.0

# Optional but recommended
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0